fix: Modernize Azure storage backend and fix Python 3.12 compatibility
- Migrate from BlockBlobService to BlobServiceClient
- Add DefaultAzureCredential support for authentication
- Implement proper URI parsing with _extract_container_and_blob
- Improve error handling and resource cleanup
- Update dependencies in setup.py
- Fix HTTPS handler scheme resolution
- Add Python 3.12 to test matrix

+++ osaka/storage/az.py
@@ -18,8 +18,9 @@ import osaka.base
 import osaka.utils
 import osaka.storage.file
 
-from configparser import SafeConfigParser
-from azure.storage.blob import BlockBlobService
+from configparser import ConfigParser
+from azure.storage.blob import BlobServiceClient
+from azure.identity import DefaultAzureCredential
 
 """
 Azure storage connection services
@@ -36,19 +37,60 @@ class Azure(osaka.base.StorageBase):
     def __init__(self):
         """
         Constructor for the Azure object
-
         """
         self.tmpfiles = []
-        self.service = None
+        self.client = None
+        self.account_name = None
+        self.account_key = None
+        self.connection_string = None
+        self.container_name = None
 
     def connect(self, uri, params={}):
         """
         Connects to the backend
         @param uri - azures or azure uri for resource
-        @param params - optional, may contain: account_name, account_key
+        @param params - optional, may contain: account_name, account_key, connection_string
         """
         osaka.utils.LOGGER.debug("Opening Azure handler")
         self.cache = {}
+        
+        # Parse account name and key from params or environment
+        self.account_name = params.get('account_name')
+        self.account_key = params.get('account_key')
+        self.connection_string = params.get('connection_string')
+        
+        # If no credentials provided, try to use DefaultAzureCredential
+        if not (self.account_name and self.account_key) and not self.connection_string:
+            try:
+                self.client = BlobServiceClient(
+                    account_url=f"https://{self.account_name}.blob.core.windows.net",
+                    credential=DefaultAzureCredential()
+                )
+                osaka.utils.LOGGER.debug("Successfully authenticated with DefaultAzureCredential")
+            except Exception as e:
+                osaka.utils.LOGGER.warning(
+                    "Failed to authenticate with DefaultAzureCredential: {}".format(str(e))
+                )
+        
+        # If still no client, try with account key or connection string
+        if self.client is None:
+            if self.connection_string:
+                self.client = BlobServiceClient.from_connection_string(self.connection_string)
+                osaka.utils.LOGGER.debug("Connected using connection string")
+            elif self.account_name and self.account_key:
+                self.client = BlobServiceClient(
+                    account_url=f"https://{self.account_name}.blob.core.windows.net",
+                    credential={"account_name": self.account_name, "account_key": self.account_key}
+                )
+                osaka.utils.LOGGER.debug("Connected using account name and key")
+            else:
+                raise osaka.utils.OsakaException(
+                    "No valid Azure credentials provided. Need either account_name/account_key or connection_string."
+                )
+        
+        # Extract container name from URI
+        parsed = urllib.parse.urlparse(uri)
+        self.container_name = parsed.netloc.split('.')[0]  # Extract container name from hostname
         uri = re.compile("^azure").sub("http", uri)
         parsed = urllib.parse.urlparse(uri)
         session_kwargs = {}
@@ -71,7 +113,7 @@ class Azure(osaka.base.StorageBase):
             # check if ~/.azure/config exists
             azure_config_path = os.environ["HOME"] + "/.azure/config"
             if os.path.isfile(azure_config_path):
-                azure_config = SafeConfigParser()
+                azure_config = ConfigParser()
                 azure_config.read(azure_config_path)
                 session_kwargs["account_name"] = azure_config.get("storage", "account")
                 session_kwargs["account_key"] = azure_config.get("storage", "key")
@@ -80,10 +122,11 @@ class Azure(osaka.base.StorageBase):
                     "No Azure Blob Storage credentials found at " + azure_config_path
                 )
 
-        self.service = BlockBlobService(**session_kwargs)
-
-        if self.service is None:
-            raise osaka.utils.OsakaException("Failed to connect to Azure")
+    def _get_service(self):
+        """
+        For backward compatibility, returns the client
+        """
+        return self.client
 
     @staticmethod
     def getSchemes():
@@ -96,27 +139,31 @@ class Azure(osaka.base.StorageBase):
 
     def get(self, uri):
         """
-        Gets the URI (azure or azures) as a stream
+        Gets the URI (azures or azure) as a stream
         @param uri: uri to get
         """
         osaka.utils.LOGGER.debug("Getting stream from URI: {0}".format(uri))
-        container, key = osaka.utils.get_container_and_path(
-            urllib.parse.urlparse(uri).path
-        )
-        fname = "/tmp/osaka-azure-%s-%s" % (uuid4(), datetime.datetime.utcnow().strftime("%Y%m%d%H%M%S.%f"))
-        with open(fname, "w"):
-            pass
-        fh = open(fname, "r+b")
-        self.tmpfiles.append(fh)
+        container_name, blob_name = self._extract_container_and_blob(uri)
+        
+        # Create a temporary file to store the blob
+        fname = "/tmp/osaka-azure-%s-%s" % (uuid4(), datetime.datetime.now(datetime.timezone.utc).strftime("%Y%m%d%H%M%S.%f"))
         try:
-            self.service.get_blob_to_path(container, key, fname)
+            # Get a blob client
+            blob_client = self.client.get_blob_client(container=container_name, blob=blob_name)
+            
+            # Download the blob to a temporary file
+            with open(fname, "wb") as download_file:
+                download_file.write(blob_client.download_blob().readall())
+            
+            # Open the file for reading and add to tmpfiles for cleanup
+            fh = open(fname, "rb")
+            self.tmpfiles.append(fh)
+            return fh
         except Exception as e:
-            osaka.utils.LOGGER.warning(
-                "Encountered exception: {}\n{}".format(e, traceback.format_exc())
-            )
-            raise osaka.utils.OsakaFileNotFound("File {} doesn't exist.".format(uri))
-        fh.seek(0)
-        return fh
+            osaka.utils.LOGGER.error("Failed to get blob: {0}".format(str(e)))
+            if os.path.exists(fname):
+                os.remove(fname)
+            raise
 
     def put(self, stream, uri):
         """
@@ -125,15 +172,27 @@ class Azure(osaka.base.StorageBase):
         @param uri: uri to put
         """
         osaka.utils.LOGGER.debug("Putting stream to URI: {0}".format(uri))
-        container, key = osaka.utils.get_container_and_path(
-            urllib.parse.urlparse(uri).path
-        )
-        self.service.create_container(container)
+        container_name, blob_name = self._extract_container_and_blob(uri)
+        
+        # Create container if it doesn't exist
+        container_client = self.client.get_container_client(container_name)
+        if not container_client.exists():
+            container_client.create_container()
+        
+        # Create a temporary file to store the stream
         with osaka.storage.file.FileHandlerConversion(stream) as fn:
-            self.service.create_blob_from_path(container, key, fn)
-        properties = self.service.get_blob_properties(container, key)
-
-        return properties.properties.content_length
+            # Upload the file
+            with open(fn, "rb") as data:
+                blob_client = self.client.get_blob_client(container=container_name, blob=blob_name)
+                blob_client.upload_blob(data, overwrite=True)
+        
+        # Get the blob properties to return the content length
+        try:
+            blob_props = blob_client.get_blob_properties()
+            return blob_props.size
+        except Exception as e:
+            osaka.utils.LOGGER.warning("Failed to get blob properties: {0}".format(str(e)))
+            return 0
 
     def listAllChildren(self, uri):
         """
@@ -146,9 +205,9 @@ class Azure(osaka.base.StorageBase):
         if "__top__" in self.cache and uri == self.cache["__top__"]:
             return [k for k in list(self.cache.keys()) if k != "__top__"]
         parsed = urllib.parse.urlparse(uri)
-        container, key = osaka.utils.get_container_and_path(parsed.path)
-        # key in this instance is used as a prefix to be filtered out.
-        collection = self.service.list_blobs(container, key)
+        container_name, blob_name = self._extract_container_and_blob(uri)
+        # blob_name in this instance is used as a prefix to be filtered out.
+        collection = self.client.list_blobs(container_name, blob_name)
         uriBase = (
             parsed.scheme
             + "://"
@@ -158,9 +217,9 @@ class Azure(osaka.base.StorageBase):
         # Setup cache, and fill it with listings
         self.cache["__top__"] = uri
         for item in collection:
-            if not (item.name == key or item.name.startswith(key + "/") or key == ""):
+            if not (item.name == blob_name or item.name.startswith(blob_name + "/") or blob_name == ""):
                 continue
-            full = uriBase + "/" + container + "/" + item.name
+            full = uriBase + "/" + container_name + "/" + item.name
             self.cache[full] = item
             ret.append(full)
         return ret
@@ -170,9 +229,13 @@ class Azure(osaka.base.StorageBase):
         Does the URI exist?
         @param uri: uri to check
         """
-        osaka.utils.LOGGER.debug("Does URI {0} exist?".format(uri))
-        # A key exists if it has some children
-        return len(self.listAllChildren(uri)) > 0
+        try:
+            container_name, blob_name = self._extract_container_and_blob(uri)
+            blob_client = self.client.get_blob_client(container=container_name, blob=blob_name)
+            return blob_client.exists()
+        except Exception as e:
+            osaka.utils.LOGGER.error("Failed to check if blob exists: {0}".format(str(e)))
+            return False
 
     def list(self, uri):
         """
@@ -182,7 +245,7 @@ class Azure(osaka.base.StorageBase):
         depth = len(uri.rstrip("/").split("/"))
         return [
             item
-            for item in self.listAllChildren()
+            for item in self.listAllChildren(uri)
             if len(item.rstrip("/").split("/")) == (depth + 1)
         ]
 
@@ -208,16 +271,20 @@ class Azure(osaka.base.StorageBase):
         for fh in self.tmpfiles:
             try:
                 fh.close()
-            except:
-                osaka.utils.LOGGER.debug(
-                    "Failed to close temporary file-handle for: {0}".format(fh.name)
-                )
+            except Exception as e:
+                osaka.utils.LOGGER.debug("Failed to close temporary file-handle for {0}: {1}".format(fh.name, str(e)))
             try:
-                os.remove(fh.name)
-            except:
-                osaka.utils.LOGGER.debug(
-                    "Failed to remove temporary file-handle for: {0}".format(fh.name)
-                )
+                if os.path.exists(fh.name):
+                    os.remove(fh.name)
+            except Exception as e:
+                osaka.utils.LOGGER.debug("Failed to remove temporary file {0}: {1}".format(fh.name, str(e)))
+        
+        # Close the client connection
+        if hasattr(self, 'client') and self.client is not None:
+            try:
+                self.client.close()
+            except Exception as e:
+                osaka.utils.LOGGER.debug("Failed to close Azure client: {0}".format(str(e)))
 
     def size(self, uri):
         """
@@ -226,28 +293,47 @@ class Azure(osaka.base.StorageBase):
         """
         if uri in self.cache:
             return self.cache[uri].size
-        container, key = osaka.utils.get_container_and_path(
-            urllib.parse.urlparse(uri).path
-        )
-        properties = self.service.get_blob_properties(container, key).properties
-        return properties.content_length
+        container_name, blob_name = self._extract_container_and_blob(uri)
+        blob_client = self.client.get_blob_client(container=container_name, blob=blob_name)
+        properties = blob_client.get_blob_properties()
+        return properties.size
 
     def rm(self, uri):
         """
         Remove this uri from backend
         @param uri: uri to remove
         """
-        container, key = osaka.utils.get_container_and_path(
-            urllib.parse.urlparse(uri).path
-        )
-        self.service.delete_blob(container, key)
+        container_name, blob_name = self._extract_container_and_blob(uri)
+        blob_client = self.client.get_blob_client(container=container_name, blob=blob_name)
+        blob_client.delete_blob()
+
+    def _extract_container_and_blob(self, uri):
+        """
+        Extracts container and blob name from URI
+        @param uri: Azure blob storage URI
+        @return: tuple of (container_name, blob_name)
+        """
+        parsed = urllib.parse.urlparse(uri)
+        path = parsed.path.lstrip('/')
+        
+        # Handle both formats: container.blob.core.windows.net/... and azure://container/...
+        if parsed.netloc and '.blob.core.windows.net' in parsed.netloc:
+            # Format: https://container.blob.core.windows.net/blob/path
+            container = parsed.netloc.split('.')[0]
+            return container, path
+        else:
+            # Format: azure://container/blob/path
+            parts = path.split('/', 1)
+            if len(parts) == 1:
+                return parts[0], ""
+            return parts[0], parts[1]
 
     def getKeysWithPrefixURI(self, uri):
         """
         Keys with prefix of given URI
         @param uri: prefix URI
         """
-        parsed = urllib.parse.urlparse(uri)
-        container, key = osaka.utils.get_container_and_path(parsed.path)
-        collection = self.service.list_blobs(container, key)
-        return [item.name + "/" + item.key for item in collection]
+        container_name, prefix = self._extract_container_and_blob(uri)
+        container_client = self.client.get_container_client(container_name)
+        blobs = container_client.list_blobs(name_starts_with=prefix)
+        return [f"{blob.name}" for blob in blobs]
